pyproject.toml
[project]
name = "chatui-template-with-htmx"
version = "0.1.0"
description = "HTMX + FastAPI chat UI template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "jinja2==3.1.4",
  "python-multipart==0.0.9",
  "pyjwt==2.9.0",
  "pydantic-settings==2.5.2",
]

[tool.uv]
package = false
